+++ Changelog.md
@@ -1,5 +1,6 @@
 ## Upcoming Version
 - Keep screen on in camera page
+- New: Create timelapse by selecting folder of jpg images and save as video using FFmpeg
 
 ## v2026.07.08 (Beta)
 - First release